==> interesses/02_doacoes.ipynb <==
# -*- coding: utf-8
# Abraji (https://www.abraji.org.br)
# Reinaldo Chaves (reinaldo@abraji.org.br)
# Programa 02 do curso Como usar Python para descobrir uma rede de interesses
#

import pandas as pd
from pathlib import Path

# Cria arquivos com dados das doações de campanhas nas eleições de 2010, 2014 e 2018

Fonte: http://www.tse.jus.br/eleicoes/estatisticas/repositorio-de-dados-eleitorais-1/repositorio-de-dados-eleitorais

# 2010

def _reader_2010(fname):
    return pd.read_csv(fname,\
                       sep=';', \
                       encoding='latin_1', \
                       converters={'cpf_candidato': lambda x: str(x),
                                   'sequencial': lambda x: str(x),\
                                   'cpf_cnpj_doador': lambda x: str(x)}, \
                       decimal=',',\
                       header=0,\
                       names=['data_hora',
                              'sequencial',
                              'uf',
                              'partido_eleicao',
                              'numero_candidato',
                              'cargo',
                              'nome_candidato',
                              'cpf_candidato',
                              'entrega_conjunto',
                              'recibo_eleitoral',
                              'numero_documento',
                              'cpf_cnpj_doador',
                              'nome_doador',
                              'data_receita',
                              'valor_receita',
                              'tipo_receita',
                              'fonte_recurso',
                              'especie_recurso',
                              'descricao_receita'])

folder = Path("doacoes/2010")
doacoes_2010 = pd.concat([
    _reader_2010(csv)
    for csv in folder.glob("*.txt")
])

doacoes_2010.info()

doacoes_2010.head()

doacoes_2010.to_csv('resultados/doacoes_2010.csv', index=False, decimal=',')

# 2014

doacoes_2014 = pd.read_csv('doacoes/2014/receitas_candidatos_2014_brasil.txt',\
                           sep=';',\
                           encoding = 'latin_1',\
                           converters={'sequencial': lambda x: str(x),
                                       'cpf_candidato': lambda x: str(x), 
                                       'cpf_cnpj_doador': lambda x: str(x),
                                       'cpf_cnpj_doador_originario': lambda x: str(x)},\
                           decimal=',',\
                           header=0,\
                           names=['cod_eleicao',
                                 'desc_eleicao',
                                 'data_hora', 
                                 'cnpj_prestador_conta',
                                 'sequencial', 
                                 'uf',
                                 'partido_eleicao', 
                                 'numero_candidato',
                                 'cargo',
                                 'nome_candidato',
                                 'cpf_candidato',   
                                 'recibo_eleitoral',
                                 'numero_documento',
                                 'cpf_cnpj_doador',
                                 'nome_doador',
                                 'nome_doador_receita',
                                 'ue_doador',
                                 'numero_partido_doador',
                                 'numero_candidato_doador',
                                 'cod_setor_economico_doador',
                                 'setor_economico_doador',
                                 'data_receita',
                                 'valor_receita',
                                 'tipo_receita',
                                 'fonte_recurso',
                                 'especie_recurso',
                                 'descricao_receita',
                                 'cpf_cnpj_doador_originario',
                                 'nome_doador_originario',
                                 'tipo_doador_originario',
                                 'setor_economico_doador_originario',
                                 'nome_doador_originario_receita'])

doacoes_2014.info()

doacoes_2014.head()

doacoes_2014.to_csv('resultados/doacoes_2014.csv', index=False, decimal=',')

# 2018

doacoes_2018_bruto = pd.read_csv('doacoes/2018/receitas_candidatos_2018_BRASIL.csv',\
                           sep=';',\
                           encoding = 'latin_1',\
                           converters={'SQ_CANDIDATO': lambda x: str(x),
                                       'NR_CPF_CANDIDATO': lambda x: str(x), 
                                       'NR_CPF_CNPJ_DOADOR': lambda x: str(x)},\
                           decimal=',')

doacoes_2018_bruto.info()

doacoes_2018_bruto.head()

doacoes_2018 = doacoes_2018_bruto.drop(['DT_GERACAO', 
                          'HH_GERACAO',
                          'ANO_ELEICAO',
                          'CD_TIPO_ELEICAO',
                          'NM_TIPO_ELEICAO',
                          'CD_ELEICAO',
                          'DS_ELEICAO',
                          'DT_ELEICAO',
                          'ST_TURNO',
                          'TP_PRESTACAO_CONTAS',
                          'DT_PRESTACAO_CONTAS',
                          'SQ_PRESTADOR_CONTAS',
                          'SG_UE',
                          'NM_UE',
                          'NR_CNPJ_PRESTADOR_CONTA',
                          'CD_CARGO',
                          'NR_PARTIDO',
                          'NM_PARTIDO',
                          'CD_FONTE_RECEITA',
                          'DS_FONTE_RECEITA',
                          'CD_ORIGEM_RECEITA',
                          'DS_ORIGEM_RECEITA',
                          'CD_NATUREZA_RECEITA',
                          'DS_NATUREZA_RECEITA',
                          'CD_ESPECIE_RECEITA',
                          'DS_ESPECIE_RECEITA',
                          'CD_ESFERA_PARTIDARIA_DOADOR',
                          'SG_UF_DOADOR',
                          'CD_MUNICIPIO_DOADOR',
                          'NM_MUNICIPIO_DOADOR',
                          'NR_CANDIDATO_DOADOR',
                          'NR_PARTIDO_DOADOR',
                          'SG_PARTIDO_DOADOR',
                          'NM_PARTIDO_DOADOR',
                          'NR_RECIBO_DOACAO',
                          'NR_DOCUMENTO_DOACAO',
                          'SQ_RECEITA',
                        ] , axis=1)

doacoes_2018.info()

doacoes_2018.head()

doacoes_2018.to_csv('resultados/doacoes_2018.csv', index=False, decimal=',')

# Cria arquivo com candidatos em 2018

candidatos = pd.read_csv('candidatos/consulta_cand_2018_BRASIL.csv',\
                         sep=';',\
                         encoding = 'latin_1',\
                         converters={'SQ_CANDIDATO': lambda x: str(x),
                                     'NR_CPF_CANDIDATO': lambda x: str(x),
                                     'NR_TITULO_ELEITORAL_CANDIDATO': lambda x: str(x),
                                    })

candidatos = candidatos.drop(['DT_GERACAO', 
                              'SG_UE',
                                        'HH_GERACAO',
                                        'ANO_ELEICAO',
                                        'CD_TIPO_ELEICAO',
                                        'NM_TIPO_ELEICAO',
                                        'NR_TURNO', 
                                        'CD_ELEICAO',
                                        'DS_ELEICAO',
                                        'DT_ELEICAO',
                                        'TP_ABRANGENCIA', 
                                        'NM_UE',
                                        'CD_CARGO',
                                        'NM_EMAIL',
                                        'CD_SITUACAO_CANDIDATURA', 
                                        'CD_DETALHE_SITUACAO_CAND',
                                        'DS_DETALHE_SITUACAO_CAND',
                                        'TP_AGREMIACAO',
                                        'NR_PARTIDO',
                                        'NM_PARTIDO',
                                        'SQ_COLIGACAO',
                                        'NM_COLIGACAO',
                                        'DS_COMPOSICAO_COLIGACAO',
                                        'CD_NACIONALIDADE',
                                        'DS_NACIONALIDADE',
                                        'SG_UF_NASCIMENTO',  
                                        'CD_MUNICIPIO_NASCIMENTO',
                                        'NM_MUNICIPIO_NASCIMENTO', 
                                        'NR_IDADE_DATA_POSSE',      
                                        'CD_GENERO',
                                        'CD_GRAU_INSTRUCAO',      
                                        'DS_GRAU_INSTRUCAO',            
                                        'CD_ESTADO_CIVIL',            
                                        'DS_ESTADO_CIVIL',       
                                        'CD_COR_RACA',    
                                        'DS_COR_RACA',                   
                                        'CD_OCUPACAO',     
                                        'DS_OCUPACAO',              
                                        'NR_DESPESA_MAX_CAMPANHA',     
                                        'CD_SIT_TOT_TURNO' ,     
                                        'ST_REELEICAO',
                                        'ST_DECLARAR_BENS',                 
                                        'NR_PROTOCOLO_CANDIDATURA',         
                                        'NR_PROCESSO'] , axis=1)

candidatos.info()

candidatos.reset_index().head()

candidatos.to_csv('resultados/candidatos_2018.csv', index=False, decimal=',')
