==> interesses/03_compara_doacoes_empresas.ipynb <==
# -*- coding: utf-8
# Abraji (https://www.abraji.org.br)
# Reinaldo Chaves (reinaldo@abraji.org.br)
# Programa 03 do curso Como usar Python para descobrir uma rede de interesses
#

import pandas as pd
import numpy as np

# Verifica se as indústrias farmacêuticas foram doadoras de campanha em 2010, 2014 e 2018

# Aponta quais beneficiários foram eleitos em 2018

Formato de valores com decimais no float

pd.options.display.float_format = '{:,.2f}'.format

Acessa empresas farmacêuticas

farma = pd.read_csv('resultados/empresas_farm_cnae_principal.csv',\
                           sep=',',\
                           encoding = 'utf_8',\
                           converters={'cnpj': lambda x: str(x),
                                       'cnae_fiscal': lambda x: str(x)},\
                           decimal=',')

# Cria coluna com CNPJ raiz 

farma['cnpj_raiz'] = farma.cnpj.str[:8]

farma.info()

Acessa resultado de eleicoes 2018

candidatos = pd.read_csv('resultados/candidatos_2018.csv',\
                         sep=',',\
                         encoding = 'utf-8',\
                         converters={'SQ_CANDIDATO': lambda x: str(x),
                                     'NR_CPF_CANDIDATO': lambda x: str(x),
                                     'NR_TITULO_ELEITORAL_CANDIDATO': lambda x: str(x),
                                    })

candidatos.info()

candidatos.groupby("DS_SIT_TOT_TURNO").NR_CPF_CANDIDATO.count().reset_index()

# Filtra apenas os eleitos

identificador = candidatos['DS_SIT_TOT_TURNO'].isin(['ELEITO', 
                                                     'ELEITO POR MÉDIA',
                                                     'ELEITO POR QP']) 

eleitos = candidatos[identificador]

eleitos.info()

Carrega doacoes

2010

doacoes_2010 = pd.read_csv('resultados/doacoes_2010.csv',\
                           sep=',',\
                           encoding = 'utf_8',\
                           converters={'cpf_candidato': lambda x: str(x),
                                       'sequencial': lambda x: str(x),
                                       'cpf_cnpj_doador': lambda x: str(x)},
                           decimal=',')

# Cria campo CNPJ raiz e que mostra quando é doação vinda PF

def limpeza(linha):
    documento = str(linha["cpf_cnpj_doador"])
    if len(documento) > 11:
        return documento[:8]
    else:
        return "documento_cpf"

doacoes_2010['cnpj_raiz'] = doacoes_2010.apply(limpeza, axis=1)

doacoes_2010.info()

2014

doacoes_2014 = pd.read_csv('resultados/doacoes_2014.csv',\
                           sep=',',\
                           encoding = 'utf_8',\
                           converters={'sequencial': lambda x: str(x),
                                       'cpf_candidato': lambda x: str(x),
                                       'cpf_cnpj_doador': lambda x: str(x),
                                       'cpf_cnpj_doador_originario': lambda x: str(x)},\
                           decimal=',')

# Cria campo CNPJ raiz e que mostra quando é doação vinda PF
# Também nos doadores originários

def limpeza1(linha):
    documento = str(linha["cpf_cnpj_doador"])
    if len(documento) > 11:
        return documento[:8]
    else:
        return "documento_cpf"

def limpeza2(linha):
    documento = str(linha["cpf_cnpj_doador_originario"])
    if len(documento) > 11:
        return documento[:8]
    else:
        return "documento_cpf"

doacoes_2014['cnpj_raiz'] = doacoes_2014.apply(limpeza1, axis=1)
doacoes_2014['cnpj_raiz_originario'] = doacoes_2014.apply(limpeza2, axis=1)

doacoes_2014.info()

2018

doacoes_2018 = pd.read_csv('resultados/doacoes_2018.csv',\
                           sep=',',\
                           encoding = 'utf_8',\
                           converters={'SQ_CANDIDATO': lambda x: str(x),
                                       'NR_CPF_CANDIDATO': lambda x: str(x), 
                                       'NR_CPF_CNPJ_DOADOR': lambda x: str(x)},\
                           decimal=',')

# Cria campo CNPJ raiz e que mostra quando é doação vinda PF

def limpeza(linha):
    documento = str(linha["NR_CPF_CNPJ_DOADOR"])
    if len(documento) > 11:
        return documento[:8]
    else:
        return "documento_cpf"

doacoes_2018['cnpj_raiz'] = doacoes_2018.apply(limpeza, axis=1)

doacoes_2018.info()

# Cruzamento 1 - de CNPJs - empresas e doacoes diretas 

# Em doacoes eleitorais o mesmo doador pode doar várias, por isso o CPF ou CNPJ pode se repetir

# No caso uso o drop_duplicates no CNPJ raiz das empresas porque são os primeiros 8 dígitos que caracterizam um mesmo grupo

# Por precaução uso o drop_duplicates nos eleitos, caso o TSE tenha duplicado candidatos - isso já aconteceu

2010

farma_doadores_diretos_2010 = pd.merge(doacoes_2010, farma.drop_duplicates('cnpj_raiz'), left_on='cnpj_raiz', right_on='cnpj_raiz')

print('O dataframe contém ' + str(farma_doadores_diretos_2010.shape[0]) + ' linhas')

# Uso o how left para manter as linhas quando os políticos não tiverem sido eleitos em 2018 

# As colunas ficam em branco que vem do dataframe eleitos neste caso

farma_doadores_diretos_2010 = pd.merge(farma_doadores_diretos_2010, \
                                       eleitos.drop_duplicates('NR_CPF_CANDIDATO'), \
                                       left_on='cpf_candidato', \
                                       right_on='NR_CPF_CANDIDATO',\
                                       how='left')

farma_doadores_diretos_2010.to_excel('resultados/compara/doacoes_diretas_empreas_2010.xlsx',sheet_name='Sheet1')

Maiores doadores

farma_doadores_diretos_2010.groupby("nome_doador").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Candidatos que mais receberam

farma_doadores_diretos_2010.groupby("nome_candidato").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Tabela dinâmica doadores e candidatos

tabela = pd.pivot_table(farma_doadores_diretos_2010,index=["nome_doador","nome_candidato"], values=["valor_receita"], aggfunc=np.sum)
result = tabela.sort_values('valor_receita', ascending=False).head(10)
result

2014

farma_doadores_diretos_2014 = pd.merge(doacoes_2014, farma.drop_duplicates('cnpj_raiz'), left_on='cnpj_raiz', right_on='cnpj_raiz')

print('O dataframe contém ' + str(farma_doadores_diretos_2014.shape[0]) + ' linhas')

farma_doadores_diretos_2014 = pd.merge(farma_doadores_diretos_2014, \
                                       eleitos.drop_duplicates('NR_CPF_CANDIDATO'), \
                                       left_on='cpf_candidato', \
                                       right_on='NR_CPF_CANDIDATO',\
                                       how='left')

farma_doadores_diretos_2014.to_excel('resultados/compara/doacoes_diretas_empresas_2014.xlsx',sheet_name='Sheet1')

Maiores doadores

farma_doadores_diretos_2014.groupby("nome_doador").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Candidatos que mais receberam

farma_doadores_diretos_2014.groupby("nome_candidato").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Tabela dinâmica doadores e candidatos

tabela = pd.pivot_table(farma_doadores_diretos_2014,index=["nome_doador","nome_candidato"], values=["valor_receita"], aggfunc=np.sum)
result = tabela.sort_values('valor_receita', ascending=False).head(10)
result

2018

farma_doadores_diretos_2018 = pd.merge(doacoes_2018, farma.drop_duplicates('cnpj_raiz'), left_on='cnpj_raiz', right_on='cnpj_raiz')

print('O dataframe contém ' + str(farma_doadores_diretos_2018.shape[0]) + ' linhas')

# Cruzamento 2 - de CNPJs - empresas e doacoes indiretas

2010

Não havia essa informação

2014

farma_doadores_indiretos_2014 = pd.merge(doacoes_2014, farma.drop_duplicates('cnpj_raiz'), left_on='cnpj_raiz_originario', right_on='cnpj_raiz')

print('O dataframe contém ' + str(farma_doadores_indiretos_2014.shape[0]) + ' linhas')

farma_doadores_indiretos_2014 = pd.merge(farma_doadores_indiretos_2014, \
                                       eleitos.drop_duplicates('NR_CPF_CANDIDATO'), \
                                       left_on='cpf_candidato', \
                                       right_on='NR_CPF_CANDIDATO',\
                                       how='left')

print('O dataframe contém ' + str(farma_doadores_indiretos_2014.shape[0]) + ' linhas')

farma_doadores_indiretos_2014.to_excel('resultados/compara/doacoes_indiretas_empresas_2014.xlsx',sheet_name='Sheet1')

Maiores doadores

farma_doadores_indiretos_2014.groupby("nome_doador_originario_receita").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Candidatos que mais receberam

farma_doadores_indiretos_2014.groupby("nome_candidato").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Tabela dinâmica doadores candidatos

tabela = pd.pivot_table(farma_doadores_indiretos_2014,index=["nome_doador_originario_receita","nome_candidato"], values=["valor_receita"], aggfunc=np.sum)
result = tabela.sort_values('valor_receita', ascending=False).head(10)
result

2018

Não havia essa informação
