==> interesses/04_compara_doacoes_socios.ipynb <==
# -*- coding: utf-8
# Abraji (https://www.abraji.org.br)
# Reinaldo Chaves (reinaldo@abraji.org.br)
# Programa 04 do curso Como usar Python para descobrir uma rede de interesses
#

import pandas as pd
import numpy as np
import unidecode

# Verifica se os sócios de indústrias farmacêuticas foram doadoras de campanha em 2010, 2014 e 2018

# Aponta quais beneficiários foram eleitos em 2018

Formato de valores com decimais no float

pd.options.display.float_format = '{:,.2f}'.format

Função para retirar acentos - útil para comparar nomes

def f(str):
    return (unidecode.unidecode(str))

# Carrega sócios

farma_socios = pd.read_csv('resultados/empresas_farm_com_socios.csv',\
                           sep=',',\
                           encoding = 'utf_8',\
                           converters={'cnpj': lambda x: str(x),
                                       'cnae_fiscal': lambda x: str(x),
                                       'cpf_representante_legal': lambda x: str(x),
                                       'nome_cidade_exterior': lambda x: str(x),
                                       'nm_pais': lambda x: str(x)
                                      },\
                           decimal=',')

farma_socios.info()

farma_socios.head()

O CPF tem sempre 11 dígitos

O formato do campo cnpjcpf_do_socio é: 3 asteriscos + os dígitos 456789 do CPF + 2 asteriscos

# Acessa resultado de eleicoes 2018

candidatos = pd.read_csv('resultados/candidatos_2018.csv',\
                         sep=',',\
                         encoding = 'utf-8',\
                         converters={'SQ_CANDIDATO': lambda x: str(x),
                                     'NR_CPF_CANDIDATO': lambda x: str(x),
                                     'NR_TITULO_ELEITORAL_CANDIDATO': lambda x: str(x),
                                    })

identificador = candidatos['DS_SIT_TOT_TURNO'].isin(['ELEITO', 
                                                     'ELEITO POR MÉDIA',
                                                     'ELEITO POR QP']) 

eleitos = candidatos[identificador]

# Carrega doacoes

2010

doacoes_2010 = pd.read_csv('resultados/doacoes_2010.csv',\
                           sep=',',\
                           encoding = 'utf_8',\
                           converters={'cpf_candidato': lambda x: str(x),
                                       'sequencial': lambda x: str(x),
                                       'cpf_cnpj_doador': lambda x: str(x),
                                       'nome_doador': lambda x: str(x)},
                           decimal=',')

def limpeza(linha):
    documento = str(linha["cpf_cnpj_doador"])
    if len(documento) > 11:
        return documento[:8]
    else:
        return "documento_cpf"

doacoes_2010['cnpj_raiz'] = doacoes_2010.apply(limpeza, axis=1)

# Separa dataframe apenas com doações vindas de CPFs

doacoes_2010_pf = doacoes_2010[(doacoes_2010['cnpj_raiz'] == 'documento_cpf')].copy()

# Cria nova coluna com o CPF mascarado dos doadores
# 11 dígitos
# 12345678901
# 01234567890 -> python começa contar pelo zero
# ***345678**

# List slicing
# [start:end]
# inclusive:exclusive

doacoes_2010_pf['cpf_mascarado']='***'+doacoes_2010_pf['cpf_cnpj_doador'].str[3:9]+'**'

# Retira acentos dos nomes

doacoes_2010_pf["nome_doador"] = doacoes_2010_pf["nome_doador"].apply(f)

doacoes_2010_pf.info()

doacoes_2010_pf.reset_index().tail()

Doações diretas de CPF

# Compara ao mesmo tempo nome do doador e o CPF mascarado

farma_doadores_diretos_cpf_2010 = pd.merge(doacoes_2010_pf, 
                                           farma_socios.drop_duplicates('nome_socio'), 
                                           left_on=['nome_doador', 'cpf_mascarado'], 
                                           right_on=['nome_socio', 'cnpj_cpf_do_socio'])

print('O dataframe contém ' + str(farma_doadores_diretos_cpf_2010.shape[0]) + ' linhas')

# Adiciona informações se o candidato foi eleito em 2018

farma_doadores_diretos_cpf_2010 = pd.merge(farma_doadores_diretos_cpf_2010, \
                                       eleitos.drop_duplicates('NR_CPF_CANDIDATO'), \
                                       left_on='cpf_candidato', \
                                       right_on='NR_CPF_CANDIDATO',\
                                       how='left')

print('O dataframe contém ' + str(farma_doadores_diretos_cpf_2010.shape[0]) + ' linhas')

farma_doadores_diretos_cpf_2010.to_excel('resultados/compara/doacoes_diretas_socios_empresas_2010.xlsx',sheet_name='Sheet1')

Maiores doadores

farma_doadores_diretos_cpf_2010.groupby("nome_doador").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Candidatos que mais receberam

farma_doadores_diretos_cpf_2010.groupby("nome_candidato").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Tabela dinâmica doadores candidatos empresas

tabela = pd.pivot_table(farma_doadores_diretos_cpf_2010,index=["nome_doador","nome_candidato", "razao_social"], values=["valor_receita"], aggfunc=np.sum)
result = tabela.sort_values('valor_receita', ascending=False).head(10)
result

Exemplo de seleção das empresas

nome = farma_doadores_diretos_cpf_2010[(farma_doadores_diretos_cpf_2010['nome_socio'] == 'MANOEL SALVIANO SOBRINHO')]
nome[["razao_social", "cnpj"]]

Etapas para mostrar os sócios que têm mais de uma empresa

# Separa dataframe apenas com nomes e CPF, sem repetir - keep first

selecao1 = farma_doadores_diretos_cpf_2010.drop_duplicates(['nome_socio'], keep='first')

selecao2 = selecao1[['nome_socio', 'cnpj_cpf_do_socio']].copy()

selecao2.info()

# Cruza esse dataframe com a base original que tem todos os dados

selecao3 = pd.merge(farma_socios, selecao2, left_on=['nome_socio', 'cnpj_cpf_do_socio'], right_on=['nome_socio', 'cnpj_cpf_do_socio'])

selecao3.shape

# Seleciona apenas colunas de interesse

selecao4 = selecao3[['cnpj', 'razao_social', 'nome_socio', 'cnpj_cpf_do_socio']].copy()

# Conta as aparições

selecao4['nome_socio'].value_counts()

# Exemplo das empresas de um nome

nome = selecao4[(selecao4['nome_socio'] == 'ALEXANDRE DELLA COLETTA')]
nome

selecao4.info()

selecao4.to_excel('resultados/compara/empresas_e_socios_das_empresas_de_2010.xlsx',sheet_name='Sheet1')

2014

doacoes_2014 = pd.read_csv('resultados/doacoes_2014.csv',\
                           sep=',',\
                           encoding = 'utf_8',\
                           converters={'sequencial': lambda x: str(x),
                                       'cpf_candidato': lambda x: str(x),
                                       'cpf_cnpj_doador': lambda x: str(x),
                                       'cpf_cnpj_doador_originario': lambda x: str(x)},\
                           decimal=',')

def limpeza1(linha):
    documento = str(linha["cpf_cnpj_doador"])
    if len(documento) > 11:
        return documento[:8]
    else:
        return "documento_cpf"

def limpeza2(linha):
    documento = str(linha["cpf_cnpj_doador_originario"])
    if len(documento) > 11:
        return documento[:8]
    else:
        return "documento_cpf"

doacoes_2014['cnpj_raiz'] = doacoes_2014.apply(limpeza1, axis=1)
doacoes_2014['cnpj_raiz_originario'] = doacoes_2014.apply(limpeza2, axis=1)

doacoes_2014.info()

# Retira acentos

doacoes_2014["nome_doador_receita"] = doacoes_2014["nome_doador_receita"].apply(f)
doacoes_2014["nome_doador_originario_receita"] = doacoes_2014["nome_doador_originario_receita"].apply(f)

# Separa apenas doacoes de PF

doacoes_2014_pf_direta = doacoes_2014[(doacoes_2014['cnpj_raiz'] == 'documento_cpf') & (doacoes_2014['nome_doador_originario'] == '#NULO')].copy()
doacoes_2014_pf_direta.shape

# Cria campo de CPF mascarado

doacoes_2014_pf_direta['cpf_cnpj_doador_direto_mascara']='***'+doacoes_2014_pf_direta['cpf_cnpj_doador'].str[3:9]+'**'

# Separa apenas doacoes de PF - originárias

doacoes_2014_pf_indireta = doacoes_2014[(doacoes_2014['cnpj_raiz_originario'] == 'documento_cpf') & (doacoes_2014['nome_doador_originario'] != '#NULO')].copy()
doacoes_2014_pf_indireta.shape

# Cria campo de CPF mascarado - doacao originária

doacoes_2014_pf_indireta['cpf_cnpj_doador_indireto_mascara']='***'+doacoes_2014_pf_direta['cpf_cnpj_doador_originario'].str[3:9]+'**'

Doações diretas de CPF

# Compara ao mesmo tempo nome do doador e o CPF mascarado

farma_doadores_diretos_cpf_2014 = pd.merge(doacoes_2014_pf_direta, farma_socios.drop_duplicates('nome_socio'), left_on=['nome_doador_receita', 'cpf_cnpj_doador_direto_mascara'], right_on=['nome_socio', 'cnpj_cpf_do_socio'])

print('O dataframe contém ' + str(farma_doadores_diretos_cpf_2014.shape[0]) + ' linhas')

farma_doadores_diretos_cpf_2014 = pd.merge(farma_doadores_diretos_cpf_2014, \
                                       eleitos.drop_duplicates('NR_CPF_CANDIDATO'), \
                                       left_on='cpf_candidato', \
                                       right_on='NR_CPF_CANDIDATO',\
                                       how='left')

print('O dataframe contém ' + str(farma_doadores_diretos_cpf_2014.shape[0]) + ' linhas')

farma_doadores_diretos_cpf_2014.to_excel('resultados/compara/doacoes_diretas_socios_empresas_2014.xlsx',sheet_name='Sheet1')

Maiores doadores 

farma_doadores_diretos_cpf_2014.groupby("nome_doador").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Candidatos que mais receberam

farma_doadores_diretos_cpf_2014.groupby("nome_candidato").valor_receita.sum().reset_index().sort_values("valor_receita", ascending=False).head(10)

Tabela dinâmica doadores candidatos empresas

tabela = pd.pivot_table(farma_doadores_diretos_cpf_2014,index=["nome_doador","nome_candidato", "razao_social"], values=["valor_receita"], aggfunc=np.sum)
result = tabela.sort_values('valor_receita', ascending=False).head(10)
result

teste = farma_doadores_diretos_cpf_2014[(farma_doadores_diretos_cpf_2014['nome_doador_receita'] == 'MARCELO HENRIQUE LIMIRIO GONCALVES')]

teste[['nome_doador_receita', 'cpf_cnpj_doador_direto_mascara', 'nome_socio', 'cnpj_cpf_do_socio']]

selecao1 = farma_doadores_diretos_cpf_2014.drop_duplicates(['nome_socio'], keep='first')
selecao2 = selecao1[['nome_socio', 'cnpj_cpf_do_socio']].copy()
selecao3 = pd.merge(farma_socios, selecao2, left_on=['nome_socio', 'cnpj_cpf_do_socio'], right_on=['nome_socio', 'cnpj_cpf_do_socio'])
selecao4 = selecao3[['cnpj', 'razao_social', 'nome_socio', 'cnpj_cpf_do_socio']].copy()
selecao4['nome_socio'].value_counts()

selecao4.to_excel('resultados/compara/empresas_e_socios_das_empresas_de_2014.xlsx',sheet_name='Sheet1')

Doações indiretas de CPF

farma_doadores_indiretos_cpf_2014 = pd.merge(doacoes_2014_pf_indireta, farma_socios.drop_duplicates('nome_socio'), left_on=['nome_doador_originario_receita', 'cpf_cnpj_doador_indireto_mascara'], right_on=['nome_socio', 'cnpj_cpf_do_socio'])

print('O dataframe contém ' + str(farma_doadores_indiretos_cpf_2014.shape[0]) + ' linhas')

2018

# E os mesmos passos para 2018

doacoes_2018 = pd.read_csv('resultados/doacoes_2018.csv',\
                           sep=',',\
                           encoding = 'utf_8',\
                           converters={'SQ_CANDIDATO': lambda x: str(x),
                                       'NR_CPF_CANDIDATO': lambda x: str(x), 
                                       'NR_CPF_CNPJ_DOADOR': lambda x: str(x)},\
                           decimal=',')

def limpeza(linha):
    documento = str(linha["NR_CPF_CNPJ_DOADOR"])
    if len(documento) > 11:
        return documento[:8]
    else:
        return "documento_cpf"

doacoes_2018['cnpj_raiz'] = doacoes_2018.apply(limpeza, axis=1)

doacoes_2018.info()

doacoes_2018_pf = doacoes_2018[(doacoes_2018['cnpj_raiz'] == 'documento_cpf')].copy()
doacoes_2018_pf['cpf_mascarado']='***'+doacoes_2018_pf['NR_CPF_CNPJ_DOADOR'].str[3:9]+'**'
doacoes_2018_pf["NM_DOADOR_RFB"] = doacoes_2018_pf["NM_DOADOR_RFB"].apply(f)

Doações diretas de CPF

farma_doadores_diretos_cpf_2018 = pd.merge(doacoes_2018_pf, farma_socios.drop_duplicates('nome_socio'), left_on=['NM_DOADOR_RFB', 'cpf_mascarado'], right_on=['nome_socio', 'cnpj_cpf_do_socio'])

print('O dataframe contém ' + str(farma_doadores_diretos_cpf_2018.shape[0]) + ' linhas')

farma_doadores_diretos_cpf_2018 = pd.merge(farma_doadores_diretos_cpf_2018, \
                                       eleitos.drop_duplicates('NR_CPF_CANDIDATO'), \
                                       left_on='NR_CPF_CANDIDATO', \
                                       right_on='NR_CPF_CANDIDATO',\
                                       how='left')

print('O dataframe contém ' + str(farma_doadores_diretos_cpf_2018.shape[0]) + ' linhas')

farma_doadores_diretos_cpf_2018.to_excel('resultados/compara/doacoes_diretas_socios_empresas_2018.xlsx',sheet_name='Sheet1')

Maiores doadores

farma_doadores_diretos_cpf_2018.groupby("NM_DOADOR_RFB").VR_RECEITA.sum().reset_index().sort_values("VR_RECEITA", ascending=False).head(10)

Candidatos que mais receberam

farma_doadores_diretos_cpf_2018.groupby("NM_CANDIDATO_x").VR_RECEITA.sum().reset_index().sort_values("VR_RECEITA", ascending=False).head(10)

Tabela dinâmica doadores candidatos empresa

tabela = pd.pivot_table(farma_doadores_diretos_cpf_2018,index=["NM_DOADOR_RFB","NM_CANDIDATO_x", "razao_social"], values=["VR_RECEITA"], aggfunc=np.sum)
result = tabela.sort_values('VR_RECEITA', ascending=False).head(10)
result

selecao1 = farma_doadores_diretos_cpf_2018.drop_duplicates(['nome_socio'], keep='first')
selecao2 = selecao1[['nome_socio', 'cnpj_cpf_do_socio']].copy()
selecao3 = pd.merge(farma_socios, selecao2, left_on=['nome_socio', 'cnpj_cpf_do_socio'], right_on=['nome_socio', 'cnpj_cpf_do_socio'])
selecao4 = selecao3[['cnpj', 'razao_social', 'nome_socio', 'cnpj_cpf_do_socio']].copy()
selecao4['nome_socio'].value_counts()

selecao4.to_excel('resultados/compara/empresas_e_socios_das_empresas_de_2018.xlsx',sheet_name='Sheet1')
