==> interesses/06_farma_api_senado.ipynb <==
# -*- coding: utf-8
# Abraji (https://www.abraji.org.br)
# Reinaldo Chaves (reinaldo@abraji.org.br)
# Programa 06 do curso Como usar Python para descobrir uma rede de interesses
#

import requests
import pandas as pd

Fonte: http://legis.senado.leg.br/dadosabertos/docs/ui/index.html

# Esta API pede para passar um header com o tipo de dado desejado

## É o item PesquisaMateriaService

headers = {"Accept" : "application/json"}

prefixo = 'http://legis.senado.leg.br/dadosabertos/materia/pesquisa/lista?ano='

# Cria lista vazia para receber resultados
projetos = []

# Percorre entre os anos de 2010 e 2018
for i in range(2010, 2019):
    ano = str(i)
    print("ANO: ", ano)
    
    url = prefixo + ano
    print(url)
    
    # Este é um exemplo de teste padrão em requests - pode ser usado sempre
    try:
        r = requests.get(url, headers=headers)
    except requests.exceptions.HTTPError as errh:
        print ("Http Error:",errh)
    except requests.exceptions.ConnectionError as errc:
        print ("Error Connecting:",errc) 
    except requests.exceptions.Timeout as errt:
        print ("Timeout Error:",errt)
    except requests.exceptions.RequestException as err:
        print ("OOps: Something Else",err)

    # Transforma o resultado em JSON
    projects = r.json()
    
    
    # Percorre o JSON a partir deste local
    for item in projects["PesquisaBasicaMateria"]["Materias"]["Materia"]:
        
        # Faz testes em cada campo de informação 
        try:
            AnoMateria = str(item['IdentificacaoMateria']['AnoMateria'])
        except KeyError:
            AnoMateria = None
            
        try:
            ApelidoMateria = item['DadosBasicosMateria']['ApelidoMateria']
        except KeyError:
            ApelidoMateria = None
            
        try:
            CodigoMateria = str(item['IdentificacaoMateria']['CodigoMateria'])
        except KeyError:
            CodigoMateria = None
            
        try:
            DescricaoIdentificacaoMateria = item['IdentificacaoMateria']['DescricaoIdentificacaoMateria']
        except KeyError:
            DescricaoIdentificacaoMateria = None
        
        try:
            DescricaoSubtipoMateria = item['IdentificacaoMateria']['DescricaoSubtipoMateria']
        except KeyError:
            DescricaoSubtipoMateria = None
            
        try:
            IndicadorTramitando = item['IdentificacaoMateria']['IndicadorTramitando']
        except KeyError:
            IndicadorTramitando = None
        
        try:
            NomeCasaIdentificacaoMateria = item['IdentificacaoMateria']['NomeCasaIdentificacaoMateria']
        except KeyError:
            NomeCasaIdentificacaoMateria = None
            
        try:
            NomeCasaIdentificacaoMateria = item['IdentificacaoMateria']['NomeCasaIdentificacaoMateria']
        except KeyError:
            NomeCasaIdentificacaoMateria = None
            
        try:
            NumeroMateria = str(item['IdentificacaoMateria']['NumeroMateria'])
        except KeyError:
            NumeroMateria = None
        
        try:
            DataApresentacao = item['DadosBasicosMateria']['DataApresentacao']
        except KeyError:
            DataApresentacao = None
        
        try:
            DataLeitura = item['DadosBasicosMateria']['DataLeitura']
        except KeyError:
            DataLeitura = None
            
        try:
            EmentaMateria = item['DadosBasicosMateria']['EmentaMateria'].strip()
        except KeyError:
            EmentaMateria = None
        
        try:
            ExplicacaoEmentaMateria = item['DadosBasicosMateria']['ExplicacaoEmentaMateria']
        except KeyError:
            ExplicacaoEmentaMateria = None
        
        try:
            IndicadorComplementar = item['DadosBasicosMateria']['IndicadorComplementar']
        except KeyError:
            IndicadorComplementar = None
        
        try:
            IndicadorComplementar = item['DadosBasicosMateria']['IndicadorComplementar']
        except KeyError:
            IndicadorComplementar = None
        
        try:
            DescricaoNatureza = item['DadosBasicosMateria']['NaturezaMateria']['DescricaoNatureza']
        except KeyError:
            DescricaoNatureza = None
            
        try:
            NomeAutor = item['AutoresPrincipais']['AutorPrincipal']['NomeAutor']
        except KeyError:
            NomeAutor = None
        except TypeError:
            NomeAutor = None
            
        try:
            IndicadorOutrosAutores = item['AutoresPrincipais']['AutorPrincipal']['IndicadorOutrosAutores']
        except KeyError:
            IndicadorOutrosAutores = None
        except TypeError:
            IndicadorOutrosAutores = None
            
        try:
            CodigoParlamentar = str(item['AutoresPrincipais']['AutorPrincipal']['IdentificacaoParlamentar']['CodigoParlamentar'])
        except KeyError:
            CodigoParlamentar = None
        except TypeError:
            CodigoParlamentar = None
        
        try:
            NomeParlamentar = item['AutoresPrincipais']['AutorPrincipal']['IdentificacaoParlamentar']['NomeParlamentar'] 
        except KeyError:
            NomeParlamentar = None
        except TypeError:
            NomeParlamentar = None
              
        try:
            NomeCompletoParlamentar = item['AutoresPrincipais']['AutorPrincipal']['IdentificacaoParlamentar']['NomeCompletoParlamentar']
        except KeyError:
            NomeCompletoParlamentar = None
        except TypeError:
            NomeCompletoParlamentar = None
        
        try:
            UfParlamentar = item['AutoresPrincipais']['AutorPrincipal']['IdentificacaoParlamentar']['UfParlamentar']
        except KeyError:
            UfParlamentar = None
        except TypeError:
            UfParlamentar = None
            
        try:
            DataSituacao = str(item['SituacaoAtual']['Autuacoes']['Autuacao']['Situacao']['DataSituacao'])
        except KeyError:
            DataSituacao = None
        except TypeError:
            DataSituacao = None
        
        try:
            DescricaoSituacao = item['SituacaoAtual']['Autuacoes']['Autuacao']['Situacao']['DescricaoSituacao']
        except KeyError:
            DescricaoSituacao = None
        except TypeError:
            DescricaoSituacao = None
        
        dicionario = {"AnoMateria": AnoMateria, 
                      "CodigoMateria": CodigoMateria, 
                      "DescricaoIdentificacaoMateria": DescricaoIdentificacaoMateria, 
                      "DescricaoSubtipoMateria": DescricaoSubtipoMateria, 
                      "IndicadorTramitando": IndicadorTramitando, 
                      "NomeCasaIdentificacaoMateria": NomeCasaIdentificacaoMateria, 
                      "NumeroMateria": NumeroMateria,
                      "ApelidoMateria": ApelidoMateria, 
                      "DataApresentacao": DataApresentacao, 
                      "DataLeitura": DataLeitura,
                      "EmentaMateria": EmentaMateria,
                      "ExplicacaoEmentaMateria": ExplicacaoEmentaMateria,
                      "IndicadorComplementar": IndicadorComplementar,
                      "DescricaoNatureza": DescricaoNatureza,
                      "NomeAutor": NomeAutor,
                      "IndicadorOutrosAutores": IndicadorOutrosAutores,
                      "CodigoParlamentar": CodigoParlamentar,
                      "NomeParlamentar": NomeParlamentar,
                      "NomeCompletoParlamentar": NomeCompletoParlamentar,
                      "UfParlamentar": UfParlamentar,
                      "DescricaoSituacao": DescricaoSituacao
                      }
        projetos.append(dicionario)

# Cria dataframe a partir dos dados coletados

df_projetos_api = pd.DataFrame(projetos)

df_projetos_api = df_projetos_api.drop_duplicates(['CodigoMateria'], keep='last')

df_projetos_api.info()

df_projetos_api.to_csv('resultados/projetos_senado_2010_2018.csv', index=False)

df_projetos_api.to_excel('resultados/projetos_senado_2010_2018.xlsx',sheet_name='Sheet1')

# Caso o programa seja executado aos poucos
# Pode começar aqui com o carregamento do arquivo

df_projetos_api = pd.read_csv("resultados/projetos_senado_2010_2018.csv",sep=',',\
                              encoding = 'utf-8', \
                              converters={'AnoMateria': lambda x: str(x), 
                                          'CodigoMateria': lambda x: str(x), 
                                          'CodigoParlamentar': lambda x: str(x),
                                          'IndicadorComplementar': lambda x: str(x),
                                          'IndicadorOutrosAutores': lambda x: str(x),
                                          'IndicadorTramitando': lambda x: str(x),
                                          'NumeroMateria': lambda x: str(x)
                                          })

df_projetos_api.info()

df_projetos_api.reset_index().tail()

# Deixa texto de ementas em maiúscula
# Retira linha com ementas vazias 

df_projetos_api['EmentaMateria'] = df_projetos_api['EmentaMateria'].str.upper()
df_projetos_api_final = df_projetos_api[pd.notnull(df_projetos_api['EmentaMateria'])]

Procura termos que remetem à indústria farmacêutica

str_choice = "FARMÁCIA POPULAR|ASSISTÊNCIA FARMACÊUTICA|REMÉDIO|FÁRMACOS|MEDICAMENTOS|FARMACÊUTICOS|COMPLEXO INDUSTRIAL E INOVAÇÃO EM SAÚDE|FARMACOQUÍMICO|RENAME|RELAÇÃO NACIONAL DE MEDICAMENTOS ESSENCIAIS|FITOSSANITÁRIOS|INSUMOS AGRÍCOLAS|DEFENSIVOS AGRÍCOLAS|AGROTÓXICOS|PATENTE|INPI|INSTITUTO NACIONAL DA PROPRIEDADE INDUSTRIAL|PDP|PARCERIA PARA O DESENVOLVIMENTO PRODUTIVO|SAÚDE INDÍGENA"
farma_proj = df_projetos_api_final[df_projetos_api_final['EmentaMateria'].\
                               str.contains(str_choice, na=False)]

farma_proj.info()

farma_proj.reset_index().head()

Anexa o link direto das proposicoes

link_prefixo = 'https://www25.senado.leg.br/web/atividade/materias/-/materia/'

def link(linha):
    documento = str(linha["CodigoMateria"])
    return link_prefixo + str(documento)

farma_proj["link_direto"] = farma_proj.apply(link, axis=1)

farma_proj.to_excel('resultados/autores_proposicoes_farma_senado_2010_2018.xlsx',sheet_name='Sheet1')

# Caso o programa seja executado aos poucos
# Pode começar aqui com o carregamento do arquivo

farma_proj = pd.read_excel("resultados/autores_proposicoes_farma_senado_2010_2018.xlsx",\
                           sheet_name='Sheet1',\
                           converters={'AnoMateria': lambda x: str(x), 
                                       'CodigoMateria': lambda x: str(x), 
                                       'CodigoParlamentar': lambda x: str(x),
                                       'IndicadorComplementar': lambda x: str(x),
                                       'IndicadorOutrosAutores': lambda x: str(x),
                                       'IndicadorTramitando': lambda x: str(x),
                                       'NumeroMateria': lambda x: str(x)
                                       })

farma_proj.info()

farma_proj['NomeCompletoParlamentar'].value_counts().head(20)

Seleciona linhas que indicam proposicoes com mais de um autor

mais_autores = farma_proj[(farma_proj['IndicadorOutrosAutores'] != 'Não')]

mais_autores.info()

a = len(mais_autores['CodigoMateria'].value_counts())
a

Vasculha a API para coletar os outros nomes

# É o item MateriaService 

prefixo = 'https://legis.senado.leg.br/dadosabertos/materia/'

projetos = []

for num, row in mais_autores.iterrows():
    codigo = str(row['CodigoMateria'].strip())
    DescricaoIdentificacaoMateria = row['DescricaoIdentificacaoMateria'].strip()

        
    url = prefixo + codigo
    print (url)
    
    try:
        r = requests.get(url, headers=headers)
    except requests.exceptions.HTTPError as errh:
        print ("Http Error:",errh)
    except requests.exceptions.ConnectionError as errc:
        print ("Error Connecting:",errc) 
    except requests.exceptions.Timeout as errt:
        print ("Timeout Error:",errt)
    except requests.exceptions.RequestException as err:
        print ("OOps: Something Else",err)

    projects = r.json()
    
    
    for item in projects["DetalheMateria"]["Materia"]["Autoria"]["Autor"]:
        #print(item)               
                  
        try:
            CodigoParlamentar = item['IdentificacaoParlamentar']['CodigoParlamentar']
        except KeyError:
            CodigoParlamentar = None
        except TypeError:
            CodigoParlamentar = None
            
        try:
            NomeCompletoParlamentar = item['IdentificacaoParlamentar']['NomeCompletoParlamentar']
        except KeyError:
            NomeCompletoParlamentar = None
        except TypeError:
            NomeCompletoParlamentar = None
            
        #print(CodigoParlamentar)
        
        dicionario = {"CodigoParlamentar": CodigoParlamentar, 
                      "NomeCompletoParlamentar": NomeCompletoParlamentar, 
                      "CodigoMateria": codigo,
                      "DescricaoIdentificacaoMateria": DescricaoIdentificacaoMateria
                      }
        projetos.append(dicionario)

df_mais_autores = pd.DataFrame(projetos)
df_mais_autores.info()

df_mais_autores.reset_index().head()

df_mais_autores.to_excel('resultados/mais_autores_proposicoes_farma_senado_2010_2018.xlsx',sheet_name='Sheet1')

df_mais_autores['NomeCompletoParlamentar'].value_counts().head(20)
